+++ Mask.cpp
@@ -64,7 +64,7 @@ float Mask::valueInLine(int x, int y, int x1, int y1, int x2, int y2)
 float Mask::valueInRegion(ImageObject* i, int x, int y, std::string expr)
 {
 	//TODO: probs memory leak
-	double tx, ty, r, g, b, h, s, v;
+	float tx, ty, r, g, b, h, s, v;
 
 	static std::string current = " ";
 	static float** newselection;
