+++ RgbTransform.h
@@ -10,7 +10,7 @@ public:
 	RgbTransform(ImageObject* image, AbstractMask* mask);
 	~RgbTransform();
 	std::string expr[3]; 
-	double x, y, r, g, b, h, s, v;
+	float x, y, r, g, b, h, s, v;
 	ExpressionParser parser[3];
 };
